% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidateWithMissingValues.R
\name{crossValidateWithMissingValues}
\alias{crossValidateWithMissingValues}
\title{Cross-validation for NMF}
\usage{
crossValidateWithMissingValues(
  data,
  k,
  reps = 3,
  n = 0.05,
  verbose = FALSE,
  missingValues = NULL,
  ...
)
}
\arguments{
\item{k}{array of factorization ranks to test}

\item{reps}{number of independent replicates to run}

\item{n}{fraction of values to handle as missing (default is 5\%, or \code{0.05})}

\item{verbose}{should updates be displayed when each factorization is completed}

\item{missingValues}{matrix of values to be imputed}

\item{...}{parameters to \code{RcppML::nmf}, not including \code{data} or \code{k}}
}
\value{
\code{data.frame} with class \code{nmfCrossValidate} with columns \code{rep}, \code{k}, and \code{value}
}
\description{
Find an "optimal" rank for a Non-Negative Matrix Factorization using cross-validation. Returns a \code{data.frame} with class \code{nmfCrossValidate}. Plot results using the \code{plot} class method.
}
\details{
A random speckled pattern of values is masked off during model fitting, and the mean squared error of prediction is evaluated after the model has reached the desired tolerance. The rank at which the model achieves the lowest error (best prediction accuracy) is the optimal rank.
}
\seealso{
\code{\link{nmf}}
}
