% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaskMatrix.R
\name{getMaskMatrix}
\alias{getMaskMatrix}
\title{Generate a Mask Matrix Based on Missing Values}
\usage{
getMaskMatrix(missingRows, missingColumns, dim, numRow, numCol)
}
\arguments{
\item{missingRows}{A list where each element contains row indices that have missing values.}

\item{missingColumns}{A list where each element contains column indices corresponding to missing values.}

\item{dim}{A string indicating whether to apply masking by "row" or "column".}

\item{numRow}{An integer specifying the total number of rows in the matrix.}

\item{numCol}{An integer specifying the total number of columns in the matrix.}
}
\value{
A sparse logical matrix of dimensions (\code{numRow}, \code{numCol}) with \code{TRUE} indicating missing values.
}
\description{
This function generates a binary mask matrix indicating missing values in either rows or columns.
}
\examples{
library(Matrix)
missingRows <- list(c(1, 3), c(2))
missingColumns <- list(c(2), c(3, 4))
getMaskMatrix(missingRows, missingColumns, dim = "row", numRow = 4, numCol = 4)

}
