% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaskMatrix.R
\name{getMaskVector}
\alias{getMaskVector}
\title{Generate a Mask Vector Based on Missing Values}
\usage{
getMaskVector(ind, dim, rows, cols, dimSize)
}
\arguments{
\item{ind}{An integer specifying the row or column index being checked.}

\item{dim}{A string, either "row" or "column", indicating which dimension is being processed.}

\item{rows}{A list where each element contains row indices that have missing values.}

\item{cols}{A list where each element contains column indices corresponding to missing values.}

\item{dimSize}{An integer specifying the length of the mask vector (either the number of columns or rows).}
}
\value{
A logical vector of length \code{dimSize}, where \code{TRUE} indicates a missing value.
}
\description{
This function generates a binary mask vector indicating missing values for a specific row or column.
}
\examples{
rows <- list(c(1, 3), c(2))
cols <- list(c(2), c(3, 4))
getMaskVector(1, "row", rows, cols, 4)

}
